Deskmate-Zero/requirements.txt
requests==2.31.0
tzdata==2024.1
feedparser==6.0.11
python-vlc
